+++ frontend/src/components/brand/index.ts
@@ -1,0 +1,4 @@
+export { HiredMark, type HiredMarkProps } from './HiredMark'
+export { HiredWordmark, type HiredWordmarkProps } from './HiredWordmark'
+export { HiredLockup, type HiredLockupProps } from './HiredLockup'
+export { HiredStacked, type HiredStackedProps } from './HiredStacked'